WIP: split POSIX up into true POSIX + Linux
