+++ serialx/platforms/__init__.py
@@ -2,23 +2,36 @@
 
 import sys
 
+try:
+    import termios  # noqa: F401
+except ImportError:
+    maybe_posix = False
+else:
+    maybe_posix = True
+
 if sys.platform == "win32":
     from .serial_win32 import (
         Win32Serial as Serial,
         Win32SerialTransport as SerialTransport,
         win32_list_serial_ports as list_serial_ports,
     )
-elif sys.platform == "linux" or sys.platform == "freebsd":
-    from .serial_posix import (
-        PosixSerial as Serial,
-        PosixSerialTransport as SerialTransport,
-        posix_list_serial_ports as list_serial_ports,
+elif sys.platform == "linux":
+    from .serial_linux import (
+        LinuxSerial as Serial,
+        LinuxSerialTransport as SerialTransport,
+        linux_list_serial_ports as list_serial_ports,
     )
 elif sys.platform == "darwin":
     from .serial_darwin import (
         DarwinSerial as Serial,
         DarwinSerialTransport as SerialTransport,
         darwin_list_serial_ports as list_serial_ports,
+    )
+elif maybe_posix:
+    from .serial_posix import (
+        PosixSerial as Serial,
+        PosixSerialTransport as SerialTransport,
+        posix_list_serial_ports as list_serial_ports,
     )
 else:
     raise RuntimeError(f"Unsupported platform: {sys.platform}")
